Refactor JpegReader to use a memory buffer
Instead of reading incrementally from a file, work on a memory buffer.
This makes the code cleaner, since for example we can peek at later bytes
with buf[ptr+1] instead of having to read() and then seek backwards.

Currently I only tested it working on a 'bytes' object, but I think it will
also work on mmapped files and other objects supporting the buffer
protocol.